+++ packages/back-end/src/models/SdkConnectionModel.ts
@@ -21,7 +21,6 @@ import {
 } from "../util/secrets";
 import { errorStringFromZodResult } from "../util/validation";
 import { purgeCDNCache } from "../util/fastly.util";
-import { SDKPayloadKey } from "../../types/sdk-payload";
 import { generateEncryptionKey, generateSigningKey } from "./ApiKeyModel";
 
 const sdkConnectionSchema = new mongoose.Schema({
@@ -252,15 +251,10 @@ export async function editSDKConnection(
     }
   );
 
-  const payloadKey: SDKPayloadKey = {
-    environment: connection.environment,
-    project: connection.project,
-  };
-
-  // Purge CDN if used
-  await purgeCDNCache(connection.organization, [payloadKey]);
-
   if (needsProxyUpdate) {
+    // Purge CDN if used
+    await purgeCDNCache(connection.organization, [connection.key]);
+
     await queueSingleProxyUpdate({
       ...connection,
       ...otherChanges,
